@@ -2,6 +2,7 @@ import PropTypes from 'prop-types';
 import React from 'react';
 import {
   EdgeInsetsPropType,
+  PointPropType,
   Platform,
   Animated as RNAnimated,
   requireNativeComponent,
@@ -16,10 +17,14 @@ import MapPolyline from './MapPolyline';
 import MapPolygon from './MapPolygon';
 import MapCircle from './MapCircle';
 import MapCallout from './MapCallout';
+import MapCalloutSubview from './MapCalloutSubview';
 import MapOverlay from './MapOverlay';
 import MapUrlTile from './MapUrlTile';
+import MapWMSTile from './MapWMSTile';
 import MapLocalTile from './MapLocalTile';
+import MapHeatMap from './MapHeatmap';
 import AnimatedRegion from './AnimatedRegion';
+import Geojson from './Geojson';
 import {
   contextTypes as childContextTypes,
   getAirMapName,
@@ -40,10 +45,7 @@ export const MAP_TYPES = {
   MUTEDSTANDARD: 'mutedStandard',
 };
 
-const GOOGLE_MAPS_ONLY_TYPES = [
-  MAP_TYPES.TERRAIN,
-  MAP_TYPES.NONE,
-];
+const GOOGLE_MAPS_ONLY_TYPES = [MAP_TYPES.TERRAIN, MAP_TYPES.NONE];
 
 const viewConfig = {
   uiViewClassName: 'AIR<provider>Map',
@@ -51,7 +53,6 @@ const viewConfig = {
     region: true,
   },
 };
-
 
 /**
  * Defines the map camera.
@@ -67,7 +68,6 @@ const CameraShape = PropTypes.shape({
   zoom: PropTypes.number.isRequired,
 });
 
-
 // if ViewPropTypes is not defined fall back to View.propType (to support RN < 0.44)
 const viewPropTypes = ViewPropTypes || View.propTypes;
 
@@ -78,10 +78,7 @@ const propTypes = {
    * Any value other than "google" will default to using
    * MapKit in iOS or GoogleMaps in android as the map provider.
    */
-  provider: PropTypes.oneOf([
-    'google',
-    'osmdroid',
-  ]),
+  provider: PropTypes.oneOf(['google']),
 
   /**
    * Used to style and layout the `MapView`.  See `StyleSheet.js` and
@@ -165,6 +162,14 @@ const propTypes = {
   zoomEnabled: PropTypes.bool,
 
   /**
+   * If `false` the user won't be able to double tap to zoom the map.
+   * However it will greatly decrease delay of tap gesture recognition.
+   * Default value is `true`.
+   *
+   */
+  zoomTapEnabled: PropTypes.bool,
+
+  /**
    *If `false` the user won't be able to zoom the map
    * Default value is `true`.
    *
@@ -172,7 +177,7 @@ const propTypes = {
    */
   zoomControlEnabled: PropTypes.bool,
 
-    /**
+  /**
    * If `false` the user won't be able to pinch/rotate the map.
    * Default value is `true`.
    *
@@ -241,6 +246,7 @@ const propTypes = {
    * A Boolean indicating whether the map shows scale information.
    * Default value is `false`
    *
+   * @platform ios
    */
   showsScale: PropTypes.bool,
 
@@ -400,6 +406,12 @@ const propTypes = {
   onMapReady: PropTypes.func,
 
   /**
+   * Callback that is called once all tiles have been loaded
+   * (or failed permanently) and labels have been rendered.
+   */
+  onMapLoaded: PropTypes.func,
+
+  /**
    * Callback that is called once the kml is fully loaded.
    */
   onKmlReady: PropTypes.func,
@@ -418,6 +430,11 @@ const propTypes = {
    * Callback that is called when user taps on the map.
    */
   onPress: PropTypes.func,
+
+  /**
+   * Callback that is called when user double taps on the map.
+   */
+  onDoublePress: PropTypes.func,
 
   /**
    * Callback that is called when user makes a "long press" somewhere on the map.
@@ -497,6 +514,13 @@ const propTypes = {
   kmlSrc: PropTypes.string,
 
   /**
+   * Offset Point x y for compass location.
+   *
+   * @platform ios
+   */
+  compassOffset: PointPropType,
+
+  /**
    * Callback that is called when a level is activated on a indoor building.
    */
   onIndoorLevelActivated: PropTypes.func,
@@ -525,10 +549,22 @@ class MapView extends React.Component {
     return { provider: this.props.provider };
   }
 
-  componentWillUpdate(nextProps) {
+  getSnapshotBeforeUpdate(prevProps) {
+    if (
+      this.state.isReady &&
+      this.props.customMapStyle !== prevProps.customMapStyle
+    ) {
+      this._updateStyle(this.props);
+    }
+    return this.props.region || null;
+  }
+
+  componentDidUpdate(prevProps, prevState, region) {
     const a = this.__lastRegion;
-    const b = nextProps.region;
-    if (!a || !b) return;
+    const b = region;
+    if (!a || !b) {
+      return;
+    }
     if (
       a.latitude !== b.latitude ||
       a.longitude !== b.longitude ||
@@ -542,13 +578,15 @@ class MapView extends React.Component {
   componentDidMount() {
     const { isReady } = this.state;
     if (isReady) {
-      this._updateStyle();
+      this._updateStyle(this.props);
     }
   }
 
-  _updateStyle() {
-    const { customMapStyle } = this.props;
-    this.map.setNativeProps({ customMapStyleString: JSON.stringify(customMapStyle) });
+  _updateStyle(props) {
+    const { customMapStyle } = props;
+    this.map.setNativeProps({
+      customMapStyleString: JSON.stringify(customMapStyle),
+    });
   }
 
   _onMapReady() {
@@ -558,15 +596,19 @@ class MapView extends React.Component {
     } else if (initialRegion) {
       this.map.setNativeProps({ initialRegion });
     }
-    this._updateStyle();
+    this._updateStyle(this.props);
     this.setState({ isReady: true }, () => {
-      if (onMapReady) onMapReady();
+      if (onMapReady) {
+        onMapReady();
+      }
     });
   }
 
   _onLayout(e) {
     const { layout } = e.nativeEvent;
-    if (!layout.width || !layout.height) return;
+    if (!layout.width || !layout.height) {
+      return;
+    }
     if (this.state.isReady && !this.__layoutCalled) {
       const { region, initialRegion } = this.props;
       if (region) {
@@ -617,8 +659,15 @@ class MapView extends React.Component {
   }
 
   animateToNavigation(location, bearing, angle, duration) {
-    console.warn('animateToNavigation() is deprecated, use animateCamera() instead');
-    this._runCommand('animateToNavigation', [location, bearing, angle, duration || 500]);
+    console.warn(
+      'animateToNavigation() is deprecated, use animateCamera() instead'
+    );
+    this._runCommand('animateToNavigation', [
+      location,
+      bearing,
+      angle,
+      duration || 500,
+    ]);
   }
 
   animateToRegion(region, duration) {
@@ -626,17 +675,23 @@ class MapView extends React.Component {
   }
 
   animateToCoordinate(latLng, duration) {
-    console.warn('animateToCoordinate() is deprecated, use animateCamera() instead');
+    console.warn(
+      'animateToCoordinate() is deprecated, use animateCamera() instead'
+    );
     this._runCommand('animateToCoordinate', [latLng, duration || 500]);
   }
 
   animateToBearing(bearing, duration) {
-    console.warn('animateToBearing() is deprecated, use animateCamera() instead');
+    console.warn(
+      'animateToBearing() is deprecated, use animateCamera() instead'
+    );
     this._runCommand('animateToBearing', [bearing, duration || 500]);
   }
 
   animateToViewingAngle(angle, duration) {
-    console.warn('animateToViewingAngle() is deprecated, use animateCamera() instead');
+    console.warn(
+      'animateToViewingAngle() is deprecated, use animateCamera() instead'
+    );
     this._runCommand('animateToViewingAngle', [angle, duration || 500]);
   }
 
@@ -669,7 +724,9 @@ class MapView extends React.Component {
    */
   async getMapBoundaries() {
     if (Platform.OS === 'android') {
-      return await NativeModules.AirMapModule.getMapBoundaries(this._getHandle());
+      return await NativeModules.AirMapModule.getMapBoundaries(
+        this._getHandle()
+      );
     } else if (Platform.OS === 'ios') {
       return await this._runCommand('getMapBoundaries', []);
     }
@@ -702,12 +759,14 @@ class MapView extends React.Component {
     // For the time being we support the legacy API on iOS.
     // This will be removed in a future release and only the
     // new Promise style API shall be supported.
-    if (Platform.OS === 'ios' && (arguments.length === 4)) {
-      console.warn('Old takeSnapshot API has been deprecated; will be removed in the near future'); //eslint-disable-line
-      const width = arguments[0]; // eslint-disable-line
-      const height = arguments[1]; // eslint-disable-line
-      const region = arguments[2]; // eslint-disable-line
-      const callback = arguments[3]; // eslint-disable-line
+    if (Platform.OS === 'ios' && arguments.length === 4) {
+      console.warn(
+        'Old takeSnapshot API has been deprecated; will be removed in the near future'
+      );
+      const width = arguments[0];
+      const height = arguments[1];
+      const region = arguments[2];
+      const callback = arguments[3];
       this._runCommand('takeSnapshot', [
         width || 0,
         height || 0,
@@ -729,10 +788,12 @@ class MapView extends React.Component {
       quality: args.quality || 1.0,
       result: args.result || 'file',
     };
-    if ((config.format !== 'png') &&
-        (config.format !== 'jpg')) throw new Error('Invalid format specified');
-    if ((config.result !== 'file') &&
-        (config.result !== 'base64')) throw new Error('Invalid result specified');
+    if (config.format !== 'png' && config.format !== 'jpg') {
+      throw new Error('Invalid format specified');
+    }
+    if (config.result !== 'file' && config.result !== 'base64') {
+      throw new Error('Invalid result specified');
+    }
 
     // Call native function
     if (Platform.OS === 'android') {
@@ -770,7 +831,10 @@ class MapView extends React.Component {
    */
   pointForCoordinate(coordinate) {
     if (Platform.OS === 'android') {
-      return NativeModules.AirMapModule.pointForCoordinate(this._getHandle(), coordinate);
+      return NativeModules.AirMapModule.pointForCoordinate(
+        this._getHandle(),
+        coordinate
+      );
     } else if (Platform.OS === 'ios') {
       return this._runCommand('pointForCoordinate', [coordinate]);
     }
@@ -788,11 +852,28 @@ class MapView extends React.Component {
    */
   coordinateForPoint(point) {
     if (Platform.OS === 'android') {
-      return NativeModules.AirMapModule.coordinateForPoint(this._getHandle(), point);
+      return NativeModules.AirMapModule.coordinateForPoint(
+        this._getHandle(),
+        point
+      );
     } else if (Platform.OS === 'ios') {
       return this._runCommand('coordinateForPoint', [point]);
     }
     return Promise.reject('coordinateForPoint not supported on this platform');
+  }
+
+  /**
+   * Get markers' centers and frames in user-space coordinates
+   *
+   * @param onlyVisible boolean true to include only visible markers, false to include all
+   *
+   * @return Promise Promise with { <identifier>: { point: Point, frame: Frame } }
+   */
+  getMarkersFrames(onlyVisible = false) {
+    if (Platform.OS === 'ios') {
+      return this._runCommand('getMarkersFrames', [onlyVisible]);
+    }
+    return Promise.reject('getMarkersFrames not supported on this platform');
   }
 
   /**
@@ -805,21 +886,43 @@ class MapView extends React.Component {
   boundingBoxForRegion(region) {
     return {
       northEast: {
-        latitude: region.latitude + (region.latitudeDelta / 2),
-        longitude: region.longitude + (region.longitudeDelta / 2),
+        latitude: region.latitude + region.latitudeDelta / 2,
+        longitude: region.longitude + region.longitudeDelta / 2,
       },
       southWest: {
-        latitude: region.latitude - (region.latitudeDelta / 2),
-        longitude: region.longitude - (region.longitudeDelta / 2),
+        latitude: region.latitude - region.latitudeDelta / 2,
+        longitude: region.longitude - region.longitudeDelta / 2,
       },
     };
   }
 
   _uiManagerCommand(name) {
-    return NativeModules.UIManager[getAirMapName(this.props.provider)].Commands[name];
+    const UIManager = NativeModules.UIManager;
+    const componentName = getAirMapName(this.props.provider);
+
+    if (!UIManager.getViewManagerConfig) {
+      // RN < 0.58
+      return UIManager[componentName].Commands[name];
+    }
+
+    // RN >= 0.58
+    return UIManager.getViewManagerConfig(componentName).Commands[name];
   }
 
   _mapManagerCommand(name) {
+    const managerName = getAirMapName(this.props.provider);
+
+    if (managerName === 'UI') {
+      const UIManager = NativeModules.UIManager;
+      if (!UIManager.getViewManagerConfig) {
+        // RN < 0.58
+        return UIManager[name];
+      }
+
+      // RN >= 0.58
+      return UIManager.getViewManagerConfig(name);
+    }
+
     return NativeModules[`${getAirMapName(this.props.provider)}Manager`][name];
   }
 
@@ -857,8 +960,11 @@ class MapView extends React.Component {
         onLayout: this._onLayout,
         ...this.props,
       };
-      if (Platform.OS === 'ios' && props.provider === ProviderConstants.PROVIDER_DEFAULT
-        && GOOGLE_MAPS_ONLY_TYPES.includes(props.mapType)) {
+      if (
+        Platform.OS === 'ios' &&
+        props.provider === ProviderConstants.PROVIDER_DEFAULT &&
+        GOOGLE_MAPS_ONLY_TYPES.includes(props.mapType)
+      ) {
         props.mapType = MAP_TYPES.standard;
       }
       props.handlePanDrag = !!props.onPanDrag;
@@ -877,7 +983,9 @@ class MapView extends React.Component {
     if (Platform.OS === 'android' && this.props.liteMode) {
       return (
         <AIRMapLite
-          ref={ref => { this.map = ref; }}
+          ref={ref => {
+            this.map = ref;
+          }}
           {...props}
         />
       );
@@ -887,7 +995,9 @@ class MapView extends React.Component {
 
     return (
       <AIRMap
-        ref={ref => { this.map = ref; }}
+        ref={ref => {
+          this.map = ref;
+        }}
         {...props}
       />
     );
@@ -900,15 +1010,18 @@ MapView.childContextTypes = childContextTypes;
 
 MapView.MAP_TYPES = MAP_TYPES;
 
-const nativeComponent = Component => requireNativeComponent(Component, MapView, {
-  nativeOnly: {
-    onChange: true,
-    onMapReady: true,
-    onKmlReady: true,
-    handlePanDrag: true,
-  },
-});
-const airMaps = {};
+const nativeComponent = Component =>
+  requireNativeComponent(Component, MapView, {
+    nativeOnly: {
+      onChange: true,
+      onMapReady: true,
+      onKmlReady: true,
+      handlePanDrag: true,
+    },
+  });
+const airMaps = {
+  default: nativeComponent('AIRMap'),
+};
 if (Platform.OS === 'android') {
   airMaps.google = googleMapIsInstalled ? nativeComponent('AIRMap') :
     createNotSupportedComponent('react-native-maps: google maps dependecy must be added to your app build.gradle to support google provider.'); // eslint-disable-line max-len
@@ -916,24 +1029,44 @@ if (Platform.OS === 'android') {
     createNotSupportedComponent('react-native-maps: osmdroid dependecy must be added to your app build.gradle to support osmdroid provider.'); // eslint-disable-line max-len
   airMaps.default = airMaps.osmdroid;
 } else {
-  airMaps.default = nativeComponent('AIRMap');
-  airMaps.google = googleMapIsInstalled ? nativeComponent('AIRGoogleMap') :
-    createNotSupportedComponent('react-native-maps: AirGoogleMaps dir must be added to your xCode project to support GoogleMaps on iOS.'); // eslint-disable-line max-len
+  airMaps.google = googleMapIsInstalled
+    ? nativeComponent('AIRGoogleMap')
+    : createNotSupportedComponent(
+        'react-native-maps: AirGoogleMaps dir must be added to your xCode project to support GoogleMaps on iOS.'
+      );
 }
 const getAirMapComponent = provider => airMaps[provider || 'default'];
 
-const AIRMapLite = NativeModules.UIManager.AIRMapLite &&
-  requireNativeComponent('AIRMapLite', MapView, {
-    nativeOnly: {
-      onChange: true,
-      onMapReady: true,
-      handlePanDrag: true,
-    },
-  });
+let AIRMapLite;
+if (!NativeModules.UIManager.getViewManagerConfig) {
+  // RN < 0.58
+  AIRMapLite =
+    NativeModules.UIManager.AIRMapLite &&
+    requireNativeComponent('AIRMapLite', MapView, {
+      nativeOnly: {
+        onChange: true,
+        onMapReady: true,
+        handlePanDrag: true,
+      },
+    });
+} else {
+  // RN >= 0.58
+  AIRMapLite =
+    NativeModules.UIManager.getViewManagerConfig('AIRMapLite') &&
+    requireNativeComponent('AIRMapLite', MapView, {
+      nativeOnly: {
+        onChange: true,
+        onMapReady: true,
+        handlePanDrag: true,
+      },
+    });
+}
 
 export const Animated = RNAnimated.createAnimatedComponent(MapView);
 
-export const ProviderPropType = PropTypes.oneOf(Object.values(ProviderConstants));
+export const ProviderPropType = PropTypes.oneOf(
+  Object.values(ProviderConstants)
+);
 
 /**
  * TODO:
@@ -948,13 +1081,18 @@ MapView.Polyline = MapPolyline;
 MapView.Polygon = MapPolygon;
 MapView.Circle = MapCircle;
 MapView.UrlTile = MapUrlTile;
+MapView.MapWMSTile = MapWMSTile;
 MapView.LocalTile = MapLocalTile;
+MapView.Heatmap = MapHeatMap;
 MapView.Overlay = MapOverlay;
 MapView.Callout = MapCallout;
+MapView.CalloutSubview = MapCalloutSubview;
 Object.assign(MapView, ProviderConstants);
 MapView.ProviderPropType = ProviderPropType;
 
 MapView.Animated = Animated;
 MapView.AnimatedRegion = AnimatedRegion;
+
+MapView.Geojson = Geojson;
 
 export default MapView;
