@@ -5,12 +5,17 @@ import {
   Text,
   Dimensions,
   TouchableOpacity,
+  Alert,
 } from 'react-native';
-import MapView, { Marker, Callout, ProviderPropType } from 'react-native-maps';
+import MapView, {
+  Marker,
+  Callout,
+  CalloutSubview,
+  ProviderPropType,
+} from 'react-native-maps';
 import CustomCallout from './CustomCallout';
 
 const { width, height } = Dimensions.get('window');
-
 const ASPECT_RATIO = width / height;
 const LATITUDE = 37.78825;
 const LONGITUDE = -122.4324;
@@ -23,6 +28,7 @@ class Callouts extends React.Component {
     super(props);
 
     this.state = {
+      cnt: 0,
       region: {
         latitude: LATITUDE,
         longitude: LONGITUDE,
@@ -38,14 +44,20 @@ class Callouts extends React.Component {
         },
         {
           coordinate: {
+            latitude: LATITUDE + SPACE,
+            longitude: LONGITUDE - SPACE,
+          },
+        },
+        {
+          coordinate: {
             latitude: LATITUDE,
             longitude: LONGITUDE,
           },
         },
         {
           coordinate: {
-            latitude: LATITUDE + SPACE,
-            longitude: LONGITUDE - SPACE,
+            latitude: LATITUDE,
+            longitude: LONGITUDE - SPACE / 2,
           },
         },
       ],
@@ -68,18 +80,18 @@ class Callouts extends React.Component {
           provider={this.props.provider}
           style={styles.map}
           initialRegion={region}
+          zoomTapEnabled={false}
         >
           <Marker
-            ref={ref => { this.marker1 = ref; }}
+            ref={ref => {
+              this.marker1 = ref;
+            }}
             coordinate={markers[0].coordinate}
             title="This is a native view"
-            description="Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation" // eslint-disable-line max-len
+            description="Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation"
           />
-          <Marker
-            coordinate={markers[1].coordinate}
-          >
+          <Marker coordinate={markers[1].coordinate}>
             <Callout style={styles.plainView}>
-
               <View>
                 <Text>This is a plain view</Text>
               </View>
@@ -89,13 +101,50 @@ class Callouts extends React.Component {
             coordinate={markers[2].coordinate}
             calloutOffset={{ x: -8, y: 28 }}
             calloutAnchor={{ x: 0.5, y: 0.4 }}
+            ref={ref => {
+              this.marker2 = ref;
+            }}
           >
-            <Callout tooltip style={styles.customView}>
+            <Callout
+              alphaHitTest
+              tooltip
+              onPress={e => {
+                if (
+                  e.nativeEvent.action === 'marker-inside-overlay-press' ||
+                  e.nativeEvent.action === 'callout-inside-press'
+                ) {
+                  return;
+                }
+
+                Alert.alert('callout pressed');
+              }}
+              style={styles.customView}
+            >
               <CustomCallout>
-                <Text>This is a custom callout bubble view</Text>
+                <Text>{`This is a custom callout bubble view ${
+                  this.state.cnt
+                }`}</Text>
+                <CalloutSubview
+                  onPress={() => {
+                    this.setState({ cnt: this.state.cnt + 1 }, () => {
+                      this.marker2.redrawCallout();
+                    });
+                  }}
+                  style={[styles.calloutButton]}
+                >
+                  <Text>Click me</Text>
+                </CalloutSubview>
               </CustomCallout>
             </Callout>
           </Marker>
+          <Marker
+            ref={ref => {
+              this.marker4 = ref;
+            }}
+            coordinate={markers[3].coordinate}
+            title="You can also open this callout"
+            description="by pressing on transparent area of custom callout"
+          />
         </MapView>
         <View style={styles.buttonContainer}>
           <View style={styles.bubble}>
@@ -103,10 +152,16 @@ class Callouts extends React.Component {
           </View>
         </View>
         <View style={styles.buttonContainer}>
-          <TouchableOpacity onPress={() => this.show()} style={[styles.bubble, styles.button]}>
+          <TouchableOpacity
+            onPress={() => this.show()}
+            style={[styles.bubble, styles.button]}
+          >
             <Text>Show</Text>
           </TouchableOpacity>
-          <TouchableOpacity onPress={() => this.hide()} style={[styles.bubble, styles.button]}>
+          <TouchableOpacity
+            onPress={() => this.hide()}
+            style={[styles.bubble, styles.button]}
+          >
             <Text>Hide</Text>
           </TouchableOpacity>
         </View>
@@ -122,7 +177,7 @@ Callouts.propTypes = {
 const styles = StyleSheet.create({
   customView: {
     width: 140,
-    height: 100,
+    height: 140,
   },
   plainView: {
     width: 60,
@@ -156,6 +211,16 @@ const styles = StyleSheet.create({
     flexDirection: 'row',
     marginVertical: 20,
     backgroundColor: 'transparent',
+  },
+  calloutButton: {
+    width: 'auto',
+    backgroundColor: 'rgba(255,255,255,0.7)',
+    paddingHorizontal: 6,
+    paddingVertical: 6,
+    borderRadius: 12,
+    alignItems: 'center',
+    marginHorizontal: 10,
+    marginVertical: 10,
   },
 });
 
