@@ -1,4 +1,9 @@
+/**
+ * @format
+ */
+
 import { AppRegistry } from 'react-native';
 import App from './App';
+import { name as appName } from './app.json';
 
-AppRegistry.registerComponent('AirMapsExplorer', () => App);
+AppRegistry.registerComponent(appName, () => App);
