@@ -51,16 +51,20 @@ public class AirMapOverlayManager extends ViewGroupManager<AirMapOverlay> {
     view.setZIndex(zIndex);
   }
 
-  // @ReactProp(name = "transparency", defaultFloat = 1.0f)
-  // public void setTransparency(AirMapOverlay view, float transparency) {
-  //   view.setTransparency(transparency);
-  // }
+  @ReactProp(name = "opacity", defaultFloat = 1.0f)
+  public void setOpacity(AirMapOverlay view, float opacity) {
+    view.setTransparency(1 - opacity);
+  }
 
   @ReactProp(name = "image")
   public void setImage(AirMapOverlay view, @Nullable String source) {
     view.setImage(source);
   }
 
+  @ReactProp(name = "tappable", defaultBoolean = false)
+  public void setTappable(AirMapOverlay view, boolean tapabble) {
+    view.setTappable(tapabble);
+  }
 
   @Override
   @Nullable
