@@ -1,9 +1,12 @@
-/* eslint-disable */
-import {Animated} from 'react-native';
+import { Animated } from 'react-native';
 
 const AnimatedWithChildren = Object.getPrototypeOf(Animated.ValueXY);
 if (__DEV__) {
-  if (AnimatedWithChildren.name !== 'AnimatedWithChildren') console.error('AnimatedRegion could not obtain AnimatedWithChildren base class');
+  if (AnimatedWithChildren.name !== 'AnimatedWithChildren') {
+    console.error(
+      'AnimatedRegion could not obtain AnimatedWithChildren base class'
+    );
+  }
 }
 // const __Animated = Object.getPrototypeOf(AnimatedWithChildren);
 // if (__Animated.name !== 'Animated') console.error('AnimatedRegion could not obtain Animated base class');
@@ -13,25 +16,30 @@ var _uniqueId = 1;
 export default class AnimatedMapRegion extends AnimatedWithChildren {
   constructor(valueIn) {
     super();
-    var value = valueIn || { // probably want to come up with better defaults
+    var value = valueIn || {
+      // probably want to come up with better defaults
       latitude: 0,
       longitude: 0,
       latitudeDelta: 0,
-      longitudeDelta: 0
+      longitudeDelta: 0,
     };
-    this.latitude = value.latitude instanceof Animated.Value
-      ? value.latitude
-      : new Animated.Value(value.latitude);
-    this.longitude = value.longitude instanceof Animated.Value
-      ? value.longitude
-      : new Animated.Value(value.longitude);
-    this.latitudeDelta = value.latitudeDelta instanceof Animated.Value
-      ? value.latitudeDelta
-      : new Animated.Value(value.latitudeDelta);
-    this.longitudeDelta = value.longitudeDelta instanceof Animated.Value
-      ? value.longitudeDelta
-      : new Animated.Value(value.longitudeDelta);
-    this._listeners = {};
+    this.latitude =
+      value.latitude instanceof Animated.Value
+        ? value.latitude
+        : new Animated.Value(value.latitude);
+    this.longitude =
+      value.longitude instanceof Animated.Value
+        ? value.longitude
+        : new Animated.Value(value.longitude);
+    this.latitudeDelta =
+      value.latitudeDelta instanceof Animated.Value
+        ? value.latitudeDelta
+        : new Animated.Value(value.latitudeDelta);
+    this.longitudeDelta =
+      value.longitudeDelta instanceof Animated.Value
+        ? value.longitudeDelta
+        : new Animated.Value(value.longitudeDelta);
+    this._regionListeners = {};
   }
 
   setValue(value) {
@@ -60,7 +68,7 @@ export default class AnimatedMapRegion extends AnimatedWithChildren {
       latitude: this.latitude.__getValue(),
       longitude: this.longitude.__getValue(),
       latitudeDelta: this.latitudeDelta.__getValue(),
-      longitudeDelta: this.longitudeDelta.__getValue()
+      longitudeDelta: this.longitudeDelta.__getValue(),
     };
   }
 
@@ -91,48 +99,58 @@ export default class AnimatedMapRegion extends AnimatedWithChildren {
     var jointCallback = (/*{value}*/) => {
       callback(this.__getValue());
     };
-    this._listeners[id] = {
+    this._regionListeners[id] = {
       latitude: this.latitude.addListener(jointCallback),
       longitude: this.longitude.addListener(jointCallback),
       latitudeDelta: this.latitudeDelta.addListener(jointCallback),
-      longitudeDelta: this.longitudeDelta.addListener(jointCallback)
+      longitudeDelta: this.longitudeDelta.addListener(jointCallback),
     };
     return id;
   }
 
   removeListener(id) {
-    this.latitude.removeListener(this._listeners[id].latitude);
-    this.longitude.removeListener(this._listeners[id].longitude);
-    this.latitudeDelta.removeListener(this._listeners[id].latitudeDelta);
-    this.longitudeDelta.removeListener(this._listeners[id].longitudeDelta);
-    delete this._listeners[id];
+    this.latitude.removeListener(this._regionListeners[id].latitude);
+    this.longitude.removeListener(this._regionListeners[id].longitude);
+    this.latitudeDelta.removeListener(this._regionListeners[id].latitudeDelta);
+    this.longitudeDelta.removeListener(
+      this._regionListeners[id].longitudeDelta
+    );
+    delete this._regionListeners[id];
   }
 
   spring(config) {
     var animations = [];
     config.hasOwnProperty('latitude') &&
-    animations.push(Animated.spring(this.latitude, {
-      ...config,
-      toValue: config.latitude
-    }));
+      animations.push(
+        Animated.spring(this.latitude, {
+          ...config,
+          toValue: config.latitude,
+        })
+      );
 
     config.hasOwnProperty('longitude') &&
-    animations.push(Animated.spring(this.longitude, {
-      ...config,
-      toValue: config.longitude
-    }));
+      animations.push(
+        Animated.spring(this.longitude, {
+          ...config,
+          toValue: config.longitude,
+        })
+      );
 
     config.hasOwnProperty('latitudeDelta') &&
-    animations.push(Animated.spring(this.latitudeDelta, {
-      ...config,
-      toValue: config.latitudeDelta
-    }));
+      animations.push(
+        Animated.spring(this.latitudeDelta, {
+          ...config,
+          toValue: config.latitudeDelta,
+        })
+      );
 
     config.hasOwnProperty('longitudeDelta') &&
-    animations.push(Animated.spring(this.longitudeDelta, {
-      ...config,
-      toValue: config.longitudeDelta
-    }));
+      animations.push(
+        Animated.spring(this.longitudeDelta, {
+          ...config,
+          toValue: config.longitudeDelta,
+        })
+      );
 
     return Animated.parallel(animations);
   }
@@ -140,28 +158,36 @@ export default class AnimatedMapRegion extends AnimatedWithChildren {
   timing(config) {
     var animations = [];
     config.hasOwnProperty('latitude') &&
-    animations.push(Animated.timing(this.latitude, {
-      ...config,
-      toValue: config.latitude
-    }));
+      animations.push(
+        Animated.timing(this.latitude, {
+          ...config,
+          toValue: config.latitude,
+        })
+      );
 
     config.hasOwnProperty('longitude') &&
-    animations.push(Animated.timing(this.longitude, {
-      ...config,
-      toValue: config.longitude
-    }));
+      animations.push(
+        Animated.timing(this.longitude, {
+          ...config,
+          toValue: config.longitude,
+        })
+      );
 
     config.hasOwnProperty('latitudeDelta') &&
-    animations.push(Animated.timing(this.latitudeDelta, {
-      ...config,
-      toValue: config.latitudeDelta
-    }));
+      animations.push(
+        Animated.timing(this.latitudeDelta, {
+          ...config,
+          toValue: config.latitudeDelta,
+        })
+      );
 
     config.hasOwnProperty('longitudeDelta') &&
-    animations.push(Animated.timing(this.longitudeDelta, {
-      ...config,
-      toValue: config.longitudeDelta
-    }));
+      animations.push(
+        Animated.timing(this.longitudeDelta, {
+          ...config,
+          toValue: config.longitudeDelta,
+        })
+      );
 
     return Animated.parallel(animations);
   }
