@@ -1,10 +1,6 @@
 import PropTypes from 'prop-types';
 import React from 'react';
-import {
-  StyleSheet,
-  ViewPropTypes,
-  View,
-} from 'react-native';
+import { StyleSheet, ViewPropTypes, View } from 'react-native';
 import decorateMapComponent, {
   SUPPORTED,
   USES_DEFAULT_IMPLEMENTATION,
@@ -17,16 +13,23 @@ const propTypes = {
   ...viewPropTypes,
   tooltip: PropTypes.bool,
   onPress: PropTypes.func,
+  alphaHitTest: PropTypes.bool,
 };
 
 const defaultProps = {
   tooltip: false,
+  alphaHitTest: false,
 };
 
 class MapCallout extends React.Component {
   render() {
     const AIRMapCallout = this.getAirComponent();
-    return <AIRMapCallout {...this.props} style={[styles.callout, this.props.style]} />;
+    return (
+      <AIRMapCallout
+        {...this.props}
+        style={[styles.callout, this.props.style]}
+      />
+    );
   }
 }
 
