@@ -49,6 +49,11 @@ const propTypes = {
   description: PropTypes.string,
 
   /**
+   * Test ID for end to end test automation
+   */
+  testID: PropTypes.string,
+
+  /**
    * A custom image to be used as the marker's icon. Only local image resources are allowed to be
    * used.
    */
@@ -72,6 +77,12 @@ const propTypes = {
    * which can be customized by this color. Ignored if a custom marker is being used.
    */
   pinColor: ColorPropType,
+
+  /**
+   * When true, the marker will be pre-selected. Setting this to true allows the user to
+   * drag the marker without needing to tap on it once to focus on it.
+   */
+  isPreselected: PropTypes.bool,
 
   /**
    * The coordinate for the marker.
@@ -251,6 +262,7 @@ class MapMarker extends React.Component {
 
     this.showCallout = this.showCallout.bind(this);
     this.hideCallout = this.hideCallout.bind(this);
+    this.redrawCallout = this.redrawCallout.bind(this);
     this.animateMarkerToCoordinate = this.animateMarkerToCoordinate.bind(this);
   }
 
@@ -266,8 +278,15 @@ class MapMarker extends React.Component {
     this._runCommand('hideCallout', []);
   }
 
+  redrawCallout() {
+    this._runCommand('redrawCallout', []);
+  }
+
   animateMarkerToCoordinate(coordinate, duration) {
-    this._runCommand('animateMarkerToCoordinate', [coordinate, duration || 500]);
+    this._runCommand('animateMarkerToCoordinate', [
+      coordinate,
+      duration || 500,
+    ]);
   }
 
   redraw() {
