@@ -1,11 +1,6 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import {
-  View,
-  StyleSheet,
-  Image,
-  Animated,
-} from 'react-native';
+import { View, StyleSheet, Image, Animated } from 'react-native';
 
 import decorateMapComponent, {
   SUPPORTED,
@@ -25,15 +20,24 @@ const propTypes = {
   image: PropTypes.any.isRequired,
   // Top left and bottom right coordinates for the overlay
   bounds: PropTypes.arrayOf(PropTypes.array.isRequired).isRequired,
+  /* Boolean to allow an overlay to be tappable and use the
+   * onPress function
+   */
+  tappable: PropTypes.bool,
+  // Callback that is called when the user presses on the overlay
+  onPress: PropTypes.func,
+  // The opacity of the overlay.
+  opacity: PropTypes.number,
 };
 
 class MapOverlay extends Component {
-
   render() {
     let image;
     if (this.props.image) {
-      if (typeof this.props.image.startsWith === 'function'
-        && this.props.image.startsWith('http')) {
+      if (
+        typeof this.props.image.startsWith === 'function' &&
+        this.props.image.startsWith('http')
+      ) {
         image = this.props.image;
       } else {
         image = Image.resolveAssetSource(this.props.image) || {};
@@ -55,6 +59,9 @@ class MapOverlay extends Component {
 
 MapOverlay.propTypes = propTypes;
 MapOverlay.viewConfig = viewConfig;
+MapOverlay.defaultProps = {
+  opacity: 1.0,
+};
 
 const styles = StyleSheet.create({
   overlay: {
