@@ -1,10 +1,7 @@
 import PropTypes from 'prop-types';
 import React from 'react';
 
-import {
-  ViewPropTypes,
-  View,
-} from 'react-native';
+import { ViewPropTypes, View } from 'react-native';
 
 import decorateMapComponent, {
   USES_DEFAULT_IMPLEMENTATION,
@@ -56,16 +53,20 @@ const propTypes = {
    * @platform ios
    */
   tileSize: PropTypes.number,
+
+  /**
+   *
+   * Allow tiles using the TMS coordinate system (origin bottom left)
+   * to be used, and displayed at their correct coordinates
+   *
+   */
+  flipY: PropTypes.bool,
 };
 
 class MapUrlTile extends React.Component {
   render() {
     const AIRMapUrlTile = this.getAirComponent();
-    return (
-      <AIRMapUrlTile
-        {...this.props}
-      />
-    );
+    return <AIRMapUrlTile {...this.props} />;
   }
 }
 
