@@ -1,21 +1,19 @@
-import PropTypes from "prop-types";
-import { requireNativeComponent, NativeModules, Platform } from "react-native";
-import {
-  PROVIDER_DEFAULT,
-  PROVIDER_GOOGLE,
-  PROVIDER_OSMDROID
-} from "./ProviderConstants";
+import PropTypes from 'prop-types';
+import { requireNativeComponent, NativeModules, Platform } from 'react-native';
+import { PROVIDER_DEFAULT, PROVIDER_GOOGLE } from './ProviderConstants';
 
 export const SUPPORTED = "SUPPORTED";
 export const USES_DEFAULT_IMPLEMENTATION = "USES_DEFAULT_IMPLEMENTATION";
 export const NOT_SUPPORTED = "NOT_SUPPORTED";
 
 export function getAirMapName(provider) {
-  if (Platform.OS === "android" && provider === PROVIDER_GOOGLE)
-    return "AIRMap";
-  if (Platform.OS === "android") return "OsmMap";
-  if (provider === PROVIDER_GOOGLE) return "AIRGoogleMap";
-  return "AIRMap";
+  if (Platform.OS === 'android') {
+    return 'AIRMap';
+  }
+  if (provider === PROVIDER_GOOGLE) {
+    return 'AIRGoogleMap';
+  }
+  return 'AIRMap';
 }
 
 function getAirComponentName(provider, component) {
@@ -27,7 +25,7 @@ export const contextTypes = {
 };
 
 export const createNotSupportedComponent = message => () => {
-  console.error(message); // eslint-disable-line no-console
+  console.error(message);
   return null;
 };
 
@@ -44,9 +42,6 @@ function getViewManagerConfig(viewManagerName) {
 export const googleMapIsInstalled = !!getViewManagerConfig(
   getAirMapName(PROVIDER_GOOGLE)
 );
-export const osmdroidIsInstalled = !!getViewManagerConfig(
-  getAirMapName(PROVIDER_OSMDROID)
-);
 
 export default function decorateMapComponent(
   Component,
@@ -57,12 +52,13 @@ export default function decorateMapComponent(
   const getDefaultComponent = () =>
     requireNativeComponent(getAirComponentName(null, componentType), Component);
 
-  Component.contextTypes = contextTypes; // eslint-disable-line no-param-reassign
+  Component.contextTypes = contextTypes;
 
-  // eslint-disable-next-line no-param-reassign
   Component.prototype.getAirComponent = function getAirComponent() {
     const provider = this.context.provider || PROVIDER_DEFAULT;
-    if (components[provider]) return components[provider];
+    if (components[provider]) {
+      return components[provider];
+    }
 
     if (provider === PROVIDER_DEFAULT) {
       components[PROVIDER_DEFAULT] = getDefaultComponent();
@@ -75,16 +71,19 @@ export default function decorateMapComponent(
     if (platformSupport === NOT_SUPPORTED) {
       components[provider] = createNotSupportedComponent(
         `react-native-maps: ${componentName} is not supported on ${Platform.OS}`
-      ); // eslint-disable-line max-len
+      );
     } else if (platformSupport === SUPPORTED) {
-      // if (provider !== PROVIDER_GOOGLE || (Platform.OS === 'ios' && googleMapIsInstalled)) {
-      if (googleMapIsInstalled) {
+      if (
+        provider !== PROVIDER_GOOGLE ||
+        (Platform.OS === 'ios' && googleMapIsInstalled)
+      ) {
         components[provider] = requireNativeComponent(componentName, Component);
       }
     } else {
       // (platformSupport === USES_DEFAULT_IMPLEMENTATION)
-      if (!components[PROVIDER_DEFAULT])
+      if (!components[PROVIDER_DEFAULT]) {
         components[PROVIDER_DEFAULT] = getDefaultComponent();
+      }
       components[provider] = components[PROVIDER_DEFAULT];
     }
 
@@ -92,7 +91,6 @@ export default function decorateMapComponent(
   };
 
   Component.prototype.getUIManagerCommand = function getUIManagerCommand(name) {
-    // eslint-disable-line no-param-reassign,max-len
     const componentName = getAirComponentName(
       this.context.provider,
       componentType
@@ -103,7 +101,6 @@ export default function decorateMapComponent(
   Component.prototype.getMapManagerCommand = function getMapManagerCommand(
     name
   ) {
-    // eslint-disable-line no-param-reassign,max-len
     const airComponentName = `${getAirComponentName(
       this.context.provider,
       componentType
