@@ -9,7 +9,7 @@
 
 #import "AIRGoogleMapMarker.h"
 #import <GoogleMaps/GoogleMaps.h>
-#import <React/RCTImageLoader.h>
+#import <React/RCTImageLoaderProtocol.h>
 #import <React/RCTUtils.h>
 #import "AIRGMSMarker.h"
 #import "AIRGoogleMapCallout.h"
@@ -117,9 +117,9 @@ CGRect unionRect(CGRect a, CGRect b) {
 
 - (void)redraw {
   if (!_realMarker.iconView) return;
-  
+
   BOOL oldValue = _realMarker.tracksViewChanges;
-  
+
   if (oldValue == YES)
   {
     // Immediate refresh, like right now. Not waiting for next frame.
@@ -149,13 +149,52 @@ CGRect unionRect(CGRect a, CGRect b) {
   return nil;
 }
 
-- (void)didTapInfoWindowOfMarker:(AIRGMSMarker *)marker {
+- (void)didTapInfoWindowOfMarker:(AIRGMSMarker *)marker point:(CGPoint)point frame:(CGRect)frame {
   if (self.calloutView && self.calloutView.onPress) {
-    id event = @{@"action": @"marker-overlay-press",
-                 @"id": self.identifier ?: @"unknown",
-                 };
+      //todo: why not 'callout-press' ?
+    id event = @{
+               @"action": @"marker-overlay-press",
+               @"id": self.identifier ?: @"unknown",
+               @"point": @{
+                   @"x": @(point.x),
+                   @"y": @(point.y),
+                   },
+               @"frame": @{
+                   @"x": @(frame.origin.x),
+                   @"y": @(frame.origin.y),
+                   @"width": @(frame.size.width),
+                   @"height": @(frame.size.height),
+                   }
+               };
     self.calloutView.onPress(event);
   }
+}
+
+- (void)didTapInfoWindowOfMarker:(AIRGMSMarker *)marker {
+    [self didTapInfoWindowOfMarker:marker point:CGPointMake(-1, -1) frame:CGRectZero];
+}
+
+- (void)didTapInfoWindowOfMarker:(AIRGMSMarker *)marker subview:(AIRGoogleMapCalloutSubview*)subview point:(CGPoint)point frame:(CGRect)frame {
+    if (subview && subview.onPress) {
+        //todo: why not 'callout-inside-press' ?
+        id event = @{
+                   @"action": @"marker-inside-overlay-press",
+                   @"id": self.identifier ?: @"unknown",
+                   @"point": @{
+                       @"x": @(point.x),
+                       @"y": @(point.y),
+                       },
+                   @"frame": @{
+                       @"x": @(frame.origin.x),
+                       @"y": @(frame.origin.y),
+                       @"width": @(frame.size.width),
+                       @"height": @(frame.size.height),
+                       }
+                   };
+        subview.onPress(event);
+    } else {
+        [self didTapInfoWindowOfMarker:marker point:point frame:frame];
+    }
 }
 
 - (void)didBeginDraggingMarker:(AIRGMSMarker *)marker {
@@ -231,7 +270,7 @@ CGRect unionRect(CGRect a, CGRect b) {
     [self iconViewInsertSubview:_iconImageView atIndex:0];
   }
 
-  _reloadImageCancellationBlock = [_bridge.imageLoader loadImageWithURLRequest:[RCTConvert NSURLRequest:_imageSrc]
+  _reloadImageCancellationBlock = [[_bridge moduleForName:@"ImageLoader"] loadImageWithURLRequest:[RCTConvert NSURLRequest:_imageSrc]
                                                                           size:self.bounds.size
                                                                          scale:RCTScreenScale()
                                                                        clipped:YES
@@ -288,7 +327,7 @@ CGRect unionRect(CGRect a, CGRect b) {
   }
 
   _reloadImageCancellationBlock =
-  [_bridge.imageLoader loadImageWithURLRequest:[RCTConvert NSURLRequest:_iconSrc]
+  [[_bridge moduleForName:@"ImageLoader"] loadImageWithURLRequest:[RCTConvert NSURLRequest:_iconSrc]
                                           size:self.bounds.size
                                          scale:RCTScreenScale()
                                        clipped:YES
@@ -350,6 +389,14 @@ CGRect unionRect(CGRect a, CGRect b) {
 
 - (BOOL)draggable {
   return _realMarker.draggable;
+}
+
+- (void)setFlat:(BOOL)flat {
+  _realMarker.flat = flat;
+}
+
+- (BOOL)flat {
+  return _realMarker.flat;
 }
 
 - (void)setTracksViewChanges:(BOOL)tracksViewChanges {
